telegram-broadcast-bot/requirements.txt
python-telegram-bot[ext]==21.7
python-dotenv==1.0.1
aiosqlite==0.20.0
yt-dlp==2024.12.13
